predictive-maintenance/inference_service/requirements.txt
# Inference Service Dependencies

# FastAPI and ASGI server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# ML Libraries
tensorflow>=2.15.0
scikit-learn>=1.3.2
numpy>=1.24.3

# Model Registry
mlflow>=2.10.0

# Kafka (real-time prediction pipeline)
kafka-python>=2.0.2

# Database (write predictions)
psycopg2-binary>=2.9.9

# Configuration
pyyaml>=6.0.1

# Rate limiting
slowapi>=0.1.9

# Optional: Monitoring and logging
prometheus-client>=0.19.0
python-json-logger>=2.0.7
